Fix scale of percentage values in table of selected cells if necessary
Percentage values are expected to be on 0-1 scale but may be provided on
a 0-100 scale. A new step will test whether the largest value in the
column (in the entire meta data table) is larger than 1 and convert the
values to a 0-1 scale (only if automatic number formatting is active).