Allow initialization of default data set from existing object
When "Cerebro.options$crb_file_to_load" is set to the name of an
existing object in the R environment, the specified object will be
loaded as a data set on launch. This is particularly useful when hosting
Cerebro on a server in "closed" mode because the data set can be loaded
on server launch and does not need to be loaded in each user session.